Update from releng/12.2, svn revision 367578.

+++ libexec/rtld-elf/libmap.h
@@ -1,8 +1,8 @@
 /*
- * $FreeBSD: releng/11.3/libexec/rtld-elf/libmap.h 255765 2013-09-21 21:03:52Z des $
+ * $FreeBSD: releng/12.2/libexec/rtld-elf/libmap.h 344011 2019-02-11 15:02:02Z kib $
  */
 
 int	lm_init (char *);
 void	lm_fini (void);
 char *	lm_find (const char *, const char *);
-char *	lm_findn (const char *, const char *, const int);
+char *	lm_findn (const char *, const char *, const size_t);
